Optimize client-side collision detection and move collision detection methods into PhysicsEngine class.

+++ UDP Server/src/model/GameObject.java
@@ -61,7 +61,7 @@ public class GameObject implements Serializable{
 		this.size = size;
 	}
 	
-	public int getSize() {
+	public short getSize() {
 		return size;
 	}
